Implement JWT authentication

+++ backend/controllers/goalController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 const getGoals = asyncHandler (async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({ user: req.user.id})
 
   res.status(200).json(goals)
 })
@@ -18,6 +19,7 @@ const setGoals = asyncHandler (async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   })
   res.status(200).json(goal)
 })
@@ -29,6 +31,20 @@ const updateGoal =  asyncHandler (async (req, res) => {
   if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
+  }
+
+  const user = await User.findById(req.user.id)
+
+  // Check for user
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -45,6 +61,18 @@ const deleteGoal = asyncHandler (async (req, res) => {
   if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
+  }  const user = await User.findById(req.user.id)
+
+  // Check for user
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await Goal.deleteOne({ _id: req.params.id})
